Split JavaScript Class files. Use automatic FileLoader to reference.
