+++ Domain/js/1_Classes/Environment.js
@@ -1,0 +1,10 @@
+class Environment
+{
+	constructor()
+	{
+		this.name = undefined;
+		this.id = undefined;
+		this.list = [];
+	}
+	console() { console.log(this); }
+}
