+++ Domain/js/1_Classes/Request.js
@@ -1,0 +1,14 @@
+class Request
+{
+	constructor()
+	{
+		this.name = undefined;
+		this.id = undefined;
+		this.requestMethod = undefined;
+		this.url = undefined;
+		this.headers = [];
+		this.params = [];
+		this.body = undefined;
+	}
+	console() { console.log(this); }
+}
