+++ Domain/js/1_Classes/Workspace.js
@@ -1,0 +1,9 @@
+class Workspace
+{
+	constructor()
+	{
+		this.name = undefined;
+		this.id = undefined;
+	}
+	console() { console.log(this); }
+}
